Remove TODO_SMF_DOCs and create new extensions file
